Refactor setups into a component
